requirements.txt
pyyaml==6.0.2
jsonschema==4.22.0